consumers/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.10
prometheus-client==0.21.1
python-dotenv==1.1.0
pydantic==2.11.4
